requirements.txt

streamlit>=1.28.0
google-generativeai>=0.3.0
Pillow>=9.5.0
pandas>=2.0.0